requirements.txt
Django>=4.2
djangorestframework
grpcio
grpcio-tools
pyodbc
mssql-django
gunicorn
requests
djangorestframework-simplejwt
channels
channels-redis